małe zmiany
