+++ Assets/Logic/Checker.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: f8b01a4fb6dae2d41a59c1d1d8b50d07
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
