+++ Assets/Logic/Enums.cs
@@ -1,0 +1,24 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Assets.Logic
+{
+    public enum PlayerType
+    {
+        Human,
+        UCT,
+        RAVE,
+        AUCT,
+        Heuristic
+    }
+
+    public enum PlayerColor
+    {
+        None,
+        Red,
+        Blue
+    }
+}
