+++ Assets/Logic/Enums.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 666c610f8660acf43b62a6eceaaacc6a
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
