+++ Assets/Logic/Game.cs
@@ -1,4 +1,5 @@
-﻿using System.Collections;
+﻿using Assets.Logic;
+using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
 
@@ -70,7 +71,9 @@ public class Game //: MonoBehaviour
             for (int j = 0; j < Board.triangleSide; j++)
             {
                 players[0].checkers[checkerIndex].SetPosition (board.fields[i, j]);
+                board.fields[i, j].playerOnField = players[0].color;
                 players[1].checkers[checkerIndex].SetPosition (board.fields[Board.side - i - 1, Board.side - j - 1]);
+                board.fields[Board.side - i - 1, Board.side - j - 1].playerOnField = players[1].color;
                 checkerIndex++;
             }
         }
@@ -79,5 +82,33 @@ public class Game //: MonoBehaviour
     public void Start ()
     {
 
+    }
+
+    public PlayerColor Win ()
+    {
+        int countBlue = 0;
+        int countRed = 0;
+
+        for (int i = 0; i < Board.side; i++) // ewentualnie zrobić po triangleSide i testować też Board.side - i - 1 itd
+        {
+            for (int j = 0; j < Board.side; j++)
+            {
+                if (board.fields[i, j].playerOnField != PlayerColor.None && board.fields[i, j].fieldType != PlayerColor.None &&
+                    board.fields[i, j].playerOnField != board.fields[i, j].fieldType)
+                {
+                    if (board.fields[i, j].playerOnField == PlayerColor.Blue)
+                        countBlue++;
+                    else
+                        countRed++;
+                }
+            }
+        }
+
+        if (countBlue == checkerCount)
+            return PlayerColor.Blue;
+        else if (countRed == checkerCount)
+            return PlayerColor.Red;
+        else
+            return PlayerColor.None;
     }
 }
